Removed unnecessary files from the repository and added .gitignore.
You added some files to the git repository that usually should stay
only on your local machine. I removed them from the git repository
and added their names to the special .gitignore file to ensure that
git will ignore them next time.

Removed files:
1. IDE configuration, because different developers can use different
project configurations (for example different layout of the panels).
2. Python bytecode, because it will be regenerated every time you
run the application, so it's easier to keep these files only on
local machine to not to commit them with every code change.

Added .gitignore file. This is a special git file, that tells
git which files in the project it should ignore.
https://git-scm.com/docs/gitignore