Added requirements.txt.
As soon as your application has some external dependencies (e.g. django,
djangorestframework), you need to store somewhere the list of these
libraries to let other developers the information – what libraries they
should install locally to run your application. Otherwise I need to
guess the requirements and their versions by looking through the code.
It's something like "ok, this project uses django, and probably the
second version, because there are autogenerated comments in the
settings.py that are specific for the 2nd version of Django".

There is a convention in python community to store the list
of all requirements in the file called requirements.txt. Take a look on
this file – each line of it contains the name of the dependency and then
the version of this dependency. In this commit I set the exact required
version for each library, but there are a lot of other different ways to
configure the required version.
https://pip.pypa.io/en/stable/user_guide/#id1

This file allows you to download the project to the new environment and
just run "pip install -r requirements.txt" to install all the
dependencies of your application.